implemented middleware Authentication in users.routes

+++ src/controllers/UsersController.js
@@ -21,9 +21,9 @@ class UsersController {
 
     async update (request, response) {
         const { name, email, currentPassword, newPassword } = request.body
-        const { id } = request.params
+        const user_id = request.user.id
 
-        const user = await knex('users').where({ id }).first()
+        const user = await knex('users').where({ id: user_id }).first()
 
         if ( email ) {
             const userWithEmail = await knex('users').where({ email }).first()
@@ -47,15 +47,15 @@ class UsersController {
         
         user.name = name ? name : user.name
 
-        await knex('users').where({ id }).update({ name: user.name, email: user.email, password: user.password, updated_at: knex.fn.now() })
+        await knex('users').where({ id: user_id }).update({ name: user.name, email: user.email, password: user.password, updated_at: knex.fn.now() })
 
         return response.json()
     }
 
     async delete (request, response) {
-        const { id } = request.params
+        const user_id = request.user.id
 
-        await knex('users').where({id}).delete()
+        await knex('users').where({id: user_id}).delete()
 
         return response.json()
     }
